库存优化

+++ yudao-module-steam/yudao-module-steam-biz/src/main/java/cn/iocoder/yudao/module/steam/service/ioinvupdate/IOInvUpdateService.java
@@ -72,10 +72,10 @@ public class IOInvUpdateService {
 
 
     /**
-     * 插入库存 (对应账户库存为空)
+     * 插入库存
      */
     @Transactional
-    public void firstInsertInventory(InventoryDto inventoryDto, BindUserDO bindUserDO) {
+    public List<InvDescDO> firstInsertInventory(InventoryDto inventoryDto, BindUserDO bindUserDO) {
         // inv 表入库
         List<InvDO> invDOList = new ArrayList<>();
         for(InventoryDto.AssetsDTO assetsDTO :inventoryDto.getAssets()){
@@ -160,13 +160,19 @@ public class IOInvUpdateService {
             }
             invDescDOList.add(invDescDO);
         }
-        // 批量插入 desc
-        invDescMapper.insertBatch(invDescDOList);
+//        // 为空  批量插入 desc
+//        if((invDescMapper.selectList(new LambdaQueryWrapperX<InvDescDO>().eq(InvDescDO::getSteamId, invDescDOList.get(0).getSteamId()))).isEmpty()){
+//            invDescMapper.insertBatch(invDescDOList);
+//        }
+        // 插入desc
 
+        invDescMapper.insertBatch(invDescDOList);
+        List<InvDescDO> invDescIdList = new ArrayList<>();
         // 给inv表绑定descID
         for(InvDO item : invDOList){
             List<InvDescDO> invDescIDList = invDescMapper.selectList(new LambdaQueryWrapperX<InvDescDO>()
                     .eq(InvDescDO::getInstanceid, item.getInstanceid())
+                    .eq(InvDescDO::getSteamId, item.getSteamId())
                     .eq(InvDescDO::getClassid, item.getClassid()));
             InvDO invDescId = new InvDO();
             invDescId.setInstanceid(item.getInstanceid());
@@ -177,7 +183,10 @@ public class IOInvUpdateService {
                     .eq(InvDO::getInstanceid, invDescId.getInstanceid())
                     .eq(InvDO::getClassid, invDescId.getClassid())
                     .eq(InvDO::getSteamId,invDescId.getSteamId()));
+            invDescIdList.add(invDescIDList.get(0));
         }
+        // 返回每一条 inv 对应的 desc 详情描述信息
+        return invDescIdList;
     }
 
 
@@ -189,13 +198,26 @@ public class IOInvUpdateService {
         invDO.setSteamId(bindUserDO.getSteamId());
         invDO.setUserId(bindUserDO.getUserId());
         invDO.setBindUserId(bindUserDO.getId());
-
-        invMapper.delete(new LambdaQueryWrapperX<InvDO>()
+        // 查询
+        List<InvDO> invDOS = invMapper.selectList(new LambdaQueryWrapperX<InvDO>()
                 .eq(InvDO::getSteamId, invDO.getSteamId())
                 .eq(InvDO::getUserId, invDO.getUserId())
                 .eq(InvDO::getBindUserId, invDO.getBindUserId())
                 .eq(InvDO::getTransferStatus, "0"));
+        if (invDOS != null && !invDOS.isEmpty()){
+            List<Long> invDescIdList = new ArrayList<>();
+            for (InvDO invDO1 : invDOS) {
+                invDescIdList.add(invDO1.getInvDescId());
+            }
+            invMapper.delete(new LambdaQueryWrapperX<InvDO>()
+                    .eq(InvDO::getSteamId, invDO.getSteamId())
+                    .eq(InvDO::getUserId, invDO.getUserId())
+                    .eq(InvDO::getBindUserId, invDO.getBindUserId())
+                    .eq(InvDO::getTransferStatus, "0"));
+            invDescMapper.delete(new LambdaQueryWrapperX<InvDescDO>().in(InvDescDO::getId, invDescIdList));
+        }
     }
+
 
     /**
      *  合并库存----查询库存方法  不分页
